requirements.txt
asgiref==3.7.2
Django==4.2.9
djangorestframework==3.15.2
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.2
